% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Belgian_Magnoliopsida_sp_ssp_var_2011}
\alias{Belgian_Magnoliopsida_sp_ssp_var_2011}
\title{Belgian Demo Dataset}
\format{
\subsection{\code{Belgian_Magnoliopsida_sp_ssp_var_2011}}{

A data frame with 1,535 rows and 23 columns:
\describe{
\item{scientific_name}{The name of the taxon without the authors}
\item{authorship}{The authors of the taxon name}
}
}
}
\source{
\url{https://www.gbif.org/dataset/39653f3e-8d6b-4a94-a202-859359c164c5}
}
\usage{
Belgian_Magnoliopsida_sp_ssp_var_2011
}
\description{
This DataFrame is based on a subset of the Belgian flora published in 2011 and
downloaded from GBIF. It only contains rows flagged as Magnoliopsida and species,
subspecies or variety rank - a total of 1535 rows.
}
\details{
THIS IS UNRELIABLE DATA! It may have been changed for illustrative purposes and
should not be used in real analyses. Please visit the source for a clean version.
}
\keyword{datasets}
