% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSBI_UK_native_sp_ssp_var_2022}
\alias{BSBI_UK_native_sp_ssp_var_2022}
\title{BSBI Demo Dataset}
\format{
\subsection{\code{BSBI_UK_native_sp_ssp_var_2022}}{

A data frame with 2,954 rows and 16 columns:
\describe{
\item{taxon name}{The name of the taxon without the authors}
\item{authority}{The authors of the taxon name}
}
}
}
\source{
\url{https://bsbi.org/taxon-lists}
}
\usage{
BSBI_UK_native_sp_ssp_var_2022
}
\description{
This DataFrame is based on a subset of the Botanical Society of the British Isles
taxon list for March 2022. It only contains rows flagged as native and species,
subspecies or variety rank - a total of 2954 rows.
}
\details{
THIS IS UNRELIABLE DATA! It may have been changed for illustrative purposes and
should not be used in real analyses. Please visit the source for a clean version.
}
\keyword{datasets}
