% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{call_name_match_api}
\alias{call_name_match_api}
\title{Call the api for a name match}
\usage{
call_name_match_api(search_string = "", fallback_to_genus = FALSE)
}
\arguments{
\item{search_string}{The taxon name to be searched on}

\item{fallback_to_genus}{True if a match at genus level will do}
}
\value{
List or String or Null
}
\description{
Call the api for a name match
}
\examples{
match_name("Rhododendron ponticum L")
match_name("Rhododendron ponticum")
match_name("Rhododendron pontica")
}
