% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{Match Names}
\usage{
match(searchTerms)
}
\arguments{
\item{searchTerms}{The search terms to be looked up.}
}
\description{
Match Names
}
\examples{
match("Rhododendron ponticum")
}
