% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{report_cache_status}
\alias{report_cache_status}
\title{Report on the name cache status
recommending that it is saved}
\usage{
report_cache_status()
}
\description{
Report on the name cache status
recommending that it is saved
}
\examples{
report_cache_status()
}
