% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{wfo_clear_skips}
\alias{wfo_clear_skips}
\title{Clear skipped rows}
\usage{
wfo_clear_skips(df)
}
\arguments{
\item{df}{The data frame for the skips to be cleared from}
}
\value{
The updated data frame
}
\description{
Clear skipped rows
}
