% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{wfo_match_df_names}
\alias{wfo_match_df_names}
\title{Match names in a data frame}
\usage{
wfo_match_df_names(
  df,
  name_col,
  authors_col = NULL,
  fallback_to_genus = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{interactive}{}
}
\description{
Match names in a data frame
}
