% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{wfo_match_name}
\alias{wfo_match_name}
\title{Match a single name string against the API}
\usage{
wfo_match_name(
  search_string = "",
  fallback_to_genus = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{search_string}{The string representation of the plant name from the data.}

\item{fallback_to_genus}{If an exact match is not found then fallback to matching a genus}
}
\value{
List containing data about the matched name or null
}
\description{
Match a single name string against the API
}
\examples{
wfo_match_name("Rhododendron ponticum")
}
