% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{wfo_set_name_cache}
\alias{wfo_set_name_cache}
\title{Set the name cache to be used in this session.
Useful to prevent repeatedly looking up the same
name string in the WFO API thus increasing speed.}
\usage{
wfo_set_name_cache(name_cache)
}
\arguments{
\item{name_cache}{The new name cache}
}
\description{
Set the name cache to be used in this session.
Useful to prevent repeatedly looking up the same
name string in the WFO API thus increasing speed.
}
\examples{
my_name_cache <- wfo_get_name_cache()
... in another session ...
wfo_set_name_cache(my_name_cache)
}
