% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{wfo_stat_matches}
\alias{wfo_stat_matches}
\title{Statistics on WFO matches within a data frame.
A utility function to see what is going on.}
\usage{
wfo_stat_matches(df)
}
\arguments{
\item{df}{A data.frame containing matches in wfo_* columns.}
}
\description{
Statistics on WFO matches within a data frame.
A utility function to see what is going on.
}
\examples{
`wfo_stat_matches(df)`
}
