% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{wfo_match_name}
\alias{wfo_match_name}
\title{Match a single name string against the API}
\usage{
wfo_match_name(
  search_string = NA,
  fallback_to_genus = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{search_string}{The string representation of the plant name from the data.}

\item{fallback_to_genus}{If an exact match is not found then fallback to matching a genus}

\item{interactive}{A boolean.
If TRUE (the default) then script will stop and present a pick list if a precise match can't be found.
If FALSE then script will just continue to the next row if no unambiguous match is found.}
}
\value{
List containing data about the matched name or null
}
\description{
Match a single name string against the API
}
\examples{
`wfo_match_name("Rhododendron ponticum")`
}
